+++ android/app/src/main/java/io/github/owenyu9292/deliverymaster/MainActivity.java
@@ -1,0 +1,5 @@
+package io.github.owenyu9292.deliverymaster;
+
+import com.getcapacitor.BridgeActivity;
+
+public class MainActivity extends BridgeActivity {}